more consistent naming: DeviceMemory covers everything, KernelMemory is one child of the device memory, frames outside of kernel memory have the device flag set in order to not access them from kernel
