+++ kernel/mythos/init/mythos/init.hh
@@ -52,7 +52,7 @@ namespace init {
     PML2,
     PML3,
     PML4,
-    STATIC_MEM,
+    DEVICE_MEM,
     SCHEDULERS_START,
     CPUDRIVER = SCHEDULERS_START+256,
     INTERRUPT_CONTROL_START,
