+++ kernel/objects/common/objects/IFrame.hh
@@ -34,10 +34,10 @@ namespace mythos {
   {
   public:
     struct Info {
-      Info(uintptr_t start, size_t size, bool kernel, bool writable) : start(PhysPtr<void>(start)), size(size), kernel(kernel), writable(writable) {}
+      Info(uintptr_t start, size_t size, bool device, bool writable) : start(PhysPtr<void>(start)), size(size), device(device), writable(writable) {}
       PhysPtr<void> start;
       size_t size;
-      bool kernel;
+      bool device;
       bool writable;
     };
