+++ kernel/objects/memory-amd64/objects/FrameDataAmd64.hh
@@ -35,47 +35,11 @@
 
 namespace mythos {
 
-  /** @todo move to protocol because this is shared with the user-mode libraries */ 
-  struct FrameSize {
-    static constexpr size_t PAGE_MIN_SIZE = 1ull << 12; // 4096 bytes
-    static constexpr size_t PAGE_SIZE_SHIFT = 9; // 9 bits per page map level
-    static constexpr size_t PAGE_SIZE_FACTOR = 1ull << PAGE_SIZE_SHIFT; // 9 bits per page map level
-    static constexpr size_t FRAME_MIN_SIZE = PAGE_MIN_SIZE; // 4096 bytes
-    static constexpr size_t FRAME_SIZE_SHIFT = 1; // 1 bit increment per mythos frame, ie. all powers of two
-    static constexpr size_t FRAME_SIZE_FACTOR = 1ull << FRAME_SIZE_SHIFT; // double the size
-    static constexpr size_t FRAME_MAX_BITS = 25; // at lest 32-12=20 bits, at most what still fits into the capability data
-
-    static constexpr size_t REGION_MAX_SIZE = FRAME_MIN_SIZE * (1ull << FRAME_MAX_BITS); // 25bits for offset = 128GiB
-    // use 2048 static memory regions to cover the whole 48 bits physical address space with 25+12 bits per region
-    // this requires 80KiB of memory with 40 bytes per region, but we are working on making it smaller :)
-    static constexpr size_t STATIC_MEMORY_REGIONS = (1ull<<48)/REGION_MAX_SIZE;
-
-    constexpr static size_t logBase(size_t n, size_t base) {
-      return ( (n<base) ? 0 : 1+logBase(n/base, base));
-    }
-
-    constexpr static size_t frameBits2Size(size_t bits) {
-      return FRAME_MIN_SIZE << (bits*FRAME_SIZE_SHIFT);
-    }
-
-    constexpr static size_t frameSize2Bits(size_t size) {
-      return logBase(size/FRAME_MIN_SIZE, FRAME_SIZE_FACTOR);
-    }
-
-    constexpr static size_t pageLevel2Size(size_t level) {
-      return PAGE_MIN_SIZE << (level*PAGE_SIZE_SHIFT);
-    }
-
-    constexpr static size_t pageSize2Level(size_t size) {
-      return logBase(size/PAGE_MIN_SIZE, PAGE_SIZE_FACTOR);
-    }
-  };
-
   /** capability data for lightweight frames and for mapped frames. */
   BITFIELD_DEF(CapData, FrameData)
   typedef protocol::Frame::FrameReq FrameReq;
   BoolField<value_t, base_t, 0> writable; // can write to this memory
-  BoolField<value_t, base_t, 1> kernel;   // can be accessed by the kernel (not device memory)
+  BoolField<value_t, base_t, 1> device;   // device memory: cannot be accessed by the kernel
   UIntField<value_t, base_t, 2, 5> sizeBits;
   UIntField<value_t, base_t, 7, 25> offset; // start in its memory region
 
@@ -107,8 +71,8 @@ namespace mythos {
     auto newRange = Range<uintptr_t>::bySize(old.getStart(base) + FrameSize::FRAME_MIN_SIZE*r.offset, FrameSize::frameBits2Size(r.sizeBits));
     if (!oldRange.contains(newRange)) THROW(Error::INSUFFICIENT_RESOURCES);
     if (r.writable && !old.writable) THROW(Error::INSUFFICIENT_RESOURCES);
-    if (r.kernel && !old.kernel) THROW(Error::INSUFFICIENT_RESOURCES);
-    auto res = FrameData().writable(r.writable).kernel(r.kernel).start(base, newRange.getStart()).size(newRange.getSize());
+    if (!r.device && old.device) THROW(Error::INSUFFICIENT_RESOURCES);
+    auto res = FrameData().writable(r.writable).device(r.device).start(base, newRange.getStart()).size(newRange.getSize());
     if (derive) return self.asDerived().withData(res);
     else return self.asReference().withData(res);
   }
